pyproject.toml
[project]
name = "phis-build"
version = "2025.9.29.0"
description = "custom build system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyinstaller>=6.14.2",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
ignore-vcs = true
packages = ["phis_build"]
exclude = ["__pycache__", "*.pyc", "*.exe~", "*.syso"]

[tool.hatch.build.targets.wheel]
ignore-vcs = true
packages = ["phis_build"]
exclude = ["__pycache__", "*.pyc", "*.exe~", "*.syso"]

[project.scripts]
phis_build = "phis_build.main:main"
phis-build = "phis_build.main:main"


[tool.ruff.lint]
ignore = [
    "E722", # 不要使用裸 except（except:），应指定异常类型
    "F541", # f-string 中未使用的表达式
    "F841", # 局部变量被赋值但未使用
    # "F401", # 导入但未使用的模块
    "F811", # 变量或函数重复定义
    "E731", # 不要把 lambda 表达式赋值给变量，建议用 def
    "E402", # import 语句不在文件顶部
]

[tool.ruff.format]
quote-style = "double"
